requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==2.3.7
requests==2.31.0
oauthlib==3.2.2
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0  # For production deployment
psycopg2-binary==2.9.9 # If using PostgreSQL
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Moment
requests
oauthlib
werkzeug
